refactor: 어플리케이션 패키지 이름 변경
- 의미를 잘 나타내도록 변경
